requirements.txt
tensorflow
jax
pennylane
numpy
pandas
scikit-learn
matplotlib
Pillow